requirements.txt
pandas
numpy
seaborn
matplotlib
zipfile36
requests
python-dateutil